requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==3.1.6
djangorestframework==3.13.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
pytz==2022.1
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.3.0
tzdata==2022.1
uvicorn==0.18.2
